refactor(pipeline): extract pill-window lifecycle into submodule
Move the floating-pill window helpers (create/show/show_msg/hide, the error +
quality-gate reject variants, and the PILL_* geometry constants) into
pipeline/pill.rs. show_pill/hide_pill are re-exported pub(crate) so external
callers (main.rs, commands/recording.rs) are unchanged; the error/reject
helpers are pub(super). Verbatim move, no behavior change.

Also fixes a snippets test assertion missed by the earlier Open Flow→Verenu
rename (uppercase expected value wasn't case-matched by the rename).
